pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mfg-finance"
version = "0.1.0"
description = "Mean Field Games solver for financial market models coupling HJB and Fokker-Planck equations."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Mean Field Games Team" }
]
dependencies = [
    "numpy>=1.23",
    "scipy>=1.9",
    "matplotlib>=3.6",
    "pyyaml>=6.0",
    "tqdm>=4.64",
    "pytest>=7.0"
]

[project.urls]
Homepage = "https://github.com/example/mfg-finance"
Documentation = "https://github.com/example/mfg-finance#readme"
Issues = "https://github.com/example/mfg-finance/issues"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
